backend/requirements.txt
langchain
openai
faiss-cpu
python-dotenv
httpx
beautifulsoup4
fastapi
uvicorn
requests
paypalrestsdk
